ultrasonic and bluetooth module

+++ Atmega128/scr/ATMEGA128/HAL/LCD/LCD_Interface.h
@@ -10,7 +10,7 @@ void LCD_WriteData(uint8_t Data,uint8_t ModeType);
 void LCD_Instruction(uint8_t Instruction,uint8_t ModeType);
 void LCD_WriteString(uint8_t *String,uint8_t ModleType);
 void LCD_Go_To_XY(uint8_t x,uint8_t y);
-void LCD_Number(uint8_t ,uint8_t ModleType);
+void LCD_Number(int32_t ,uint8_t ModleType);
